Compile desugared terms.
