+++ ast/src/base.rs
@@ -1,7 +1,6 @@
 use crate::identifier::{Interner, interner, Variable, ConstructorName, FunctionName, ExternalName, Identifier};
 use crate::parser::base::PreTypeDeclaration;
 use std::collections::HashMap;
-use std::collections::HashSet;
 
 // ===Program===
 #[derive(Debug)]
@@ -379,166 +378,6 @@ impl ForeignFunctionDeclaration {
 impl ConstructorDeclaration {
     pub fn arity(&self) -> usize { 
         self.parameters.len()
-    }
-}
-
-impl Function {
-    pub fn free_variables(&self) -> Vec<Variable> {
-        let mut env: Env = Env::new();
-        free_variables_in_function(&mut env, self);
-        env.free_vars
-    }
-}
-
-// ===Free Variables===
-struct Env {
-    scope_stack: Vec<HashSet<Variable>>,
-    free_vars: Vec<Variable>,
-    free_vars_set: HashSet<Variable>,
-}
-
-impl Env {
-    fn new() -> Self {
-        Self { scope_stack: vec![], free_vars: vec![], free_vars_set: HashSet::new() }
-    }
-
-    fn open(&mut self) {
-        self.scope_stack.push(HashSet::new())
-    }
-
-    fn close(&mut self) {
-        let _ = self.scope_stack.pop();
-    }
-
-    fn extend_var(&mut self, var: &Variable) {
-        match self.scope_stack.last_mut() {
-            Some(scope) => {
-                let _ = scope.insert(var.clone());
-            },
-            None => unreachable!(),
-        }
-    }
-
-    fn extend_vars(&mut self, vars: &[Variable]) {
-        for var in vars {
-            self.extend_var(var)
-        }
-    }
-
-    fn extend_pattern(&mut self, pattern: &Pattern) { 
-        use Pattern::*;
-        match pattern {
-            Constructor(_, patterns) => {
-                for pattern in patterns {
-                    self.extend_pattern(pattern)
-                }
-            },
-            Variable(var) => {
-                self.extend_var(var)
-            },
-            Int(_) => {},
-            Anything(_) => {},
-        }
-    }
-
-    fn is_bound(&self, var: &Variable) -> bool {
-        for scope in self.scope_stack.iter().rev() {
-            if scope.contains(var) {
-                return true
-            }
-        }
-        false
-    }
-
-    fn attempt_to_register_free(&mut self, var: &Variable) {
-        if !(self.free_vars_set.contains(var) || self.is_bound(var)) {
-            self.free_vars.push(var.clone());
-            self.free_vars_set.insert(var.clone());
-        } 
-    }
-}
-
-fn free_variables_in_function(env: &mut Env, function: &Function) {
-    env.open();
-    env.extend_vars(&function.parameters);
-    free_variables(env, &function.body);
-    env.close();
-}
-
-fn free_variables(env: &mut Env, term: &Term) {
-    use Term::*;
-
-    match term {
-        TypedTerm(typed_term) => free_variables(env, &typed_term.term),
-        VariableUse(var) => {
-            env.attempt_to_register_free(var)
-        },
-        Int(_) => {}, 
-        FunctionApplication(_, _, args) => {
-            for arg in args {
-                free_variables(env, arg)
-            }
-        },
-        ConstructorUse(_, args) => {
-            for arg in args {
-                free_variables(env, arg)
-            }
-        },
-        Match(arg, branches) => {
-            free_variables(env, arg);
-            for branch in branches {
-                env.open();
-                env.extend_pattern(&branch.pattern);
-                free_variables(env, &branch.body);
-                env.close();
-            }
-        },
-        Fold(arg, branches) => {
-            free_variables(env, arg);
-            for branch in branches {
-                env.open();
-                env.extend_pattern(&branch.pattern);
-                free_variables(env, &branch.body);
-                env.close();
-            }
-        },
-        Lambda(function) => {
-            free_variables_in_function(env, function)
-        },
-        LambdaApplication(fn_term, args) => {
-            free_variables(env, fn_term);
-            for arg in args {
-                free_variables(env, arg)
-            }
-        },
-        Let(bindings, body) => {
-            env.open();
-            for (var, term) in bindings {
-                env.extend_var(var);
-                free_variables(env, term);
-            }
-            free_variables(env, body);
-            env.close();
-        },
-        Pure(term) => free_variables(env, term),
-        Do(bindings, body) => {
-            env.open();
-            for binding in bindings {
-                match binding {
-                    DoBinding::ExecuteThenBind(var, term) => {
-                        env.extend_var(var);
-                        free_variables(env, term);
-                    },
-                    DoBinding::Bind(var, term) => {
-                        env.extend_var(var);
-                        free_variables(env, term);
-                    },
-                }
-
-            }
-            free_variables(env, body);
-            env.close();
-        },
     }
 }
