+++ ast/src/lib.rs
@@ -2,3 +2,4 @@ pub mod parser;
 pub mod validation;
 pub mod identifier;
 pub mod base;
+pub mod desugared_base;
